requirements.txt
azure-functions
azure-functions-durable
agent-framework-azurefunctions
azure-identity
azure-ai-documentintelligence
dspy
pydantic
azure-storage-blob
requests
pydantic-settings
azure-monitor-opentelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-logging